Moved camera from fragment to activity. We need to redraw camera sometimes.
